+++ packages/wenyou_api/lib/src/model/moment_detail_response_dto.g.dart
@@ -203,6 +203,8 @@ class _$MomentDetailResponseDto extends MomentDetailResponseDto {
   @override
   final bool viewerBookmarked;
   @override
+  final bool? canInteract;
+  @override
   final DateTime createdAt;
   @override
   final DateTime updatedAt;
@@ -237,6 +239,7 @@ class _$MomentDetailResponseDto extends MomentDetailResponseDto {
     required this.tipTotal,
     required this.viewerLiked,
     required this.viewerBookmarked,
+    this.canInteract,
     required this.createdAt,
     required this.updatedAt,
     required this.content,
@@ -273,6 +276,7 @@ class _$MomentDetailResponseDto extends MomentDetailResponseDto {
         tipTotal == other.tipTotal &&
         viewerLiked == other.viewerLiked &&
         viewerBookmarked == other.viewerBookmarked &&
+        canInteract == other.canInteract &&
         createdAt == other.createdAt &&
         updatedAt == other.updatedAt &&
         content == other.content &&
@@ -300,6 +304,7 @@ class _$MomentDetailResponseDto extends MomentDetailResponseDto {
     _$hash = $jc(_$hash, tipTotal.hashCode);
     _$hash = $jc(_$hash, viewerLiked.hashCode);
     _$hash = $jc(_$hash, viewerBookmarked.hashCode);
+    _$hash = $jc(_$hash, canInteract.hashCode);
     _$hash = $jc(_$hash, createdAt.hashCode);
     _$hash = $jc(_$hash, updatedAt.hashCode);
     _$hash = $jc(_$hash, content.hashCode);
@@ -329,6 +334,7 @@ class _$MomentDetailResponseDto extends MomentDetailResponseDto {
           ..add('tipTotal', tipTotal)
           ..add('viewerLiked', viewerLiked)
           ..add('viewerBookmarked', viewerBookmarked)
+          ..add('canInteract', canInteract)
           ..add('createdAt', createdAt)
           ..add('updatedAt', updatedAt)
           ..add('content', content)
@@ -415,6 +421,10 @@ class MomentDetailResponseDtoBuilder
   set viewerBookmarked(bool? viewerBookmarked) =>
       _$this._viewerBookmarked = viewerBookmarked;
 
+  bool? _canInteract;
+  bool? get canInteract => _$this._canInteract;
+  set canInteract(bool? canInteract) => _$this._canInteract = canInteract;
+
   DateTime? _createdAt;
   DateTime? get createdAt => _$this._createdAt;
   set createdAt(DateTime? createdAt) => _$this._createdAt = createdAt;
@@ -467,6 +477,7 @@ class MomentDetailResponseDtoBuilder
       _tipTotal = $v.tipTotal;
       _viewerLiked = $v.viewerLiked;
       _viewerBookmarked = $v.viewerBookmarked;
+      _canInteract = $v.canInteract;
       _createdAt = $v.createdAt;
       _updatedAt = $v.updatedAt;
       _content = $v.content;
@@ -565,6 +576,7 @@ class MomentDetailResponseDtoBuilder
               r'MomentDetailResponseDto',
               'viewerBookmarked',
             ),
+            canInteract: canInteract,
             createdAt: BuiltValueNullFieldError.checkNotNull(
               createdAt,
               r'MomentDetailResponseDto',