+++ packages/wenyou_api/lib/src/model/moment_search_response_dto.g.dart
@@ -203,6 +203,8 @@ class _$MomentSearchResponseDto extends MomentSearchResponseDto {
   @override
   final bool viewerBookmarked;
   @override
+  final bool? canInteract;
+  @override
   final DateTime createdAt;
   @override
   final DateTime updatedAt;
@@ -229,6 +231,7 @@ class _$MomentSearchResponseDto extends MomentSearchResponseDto {
     required this.tipTotal,
     required this.viewerLiked,
     required this.viewerBookmarked,
+    this.canInteract,
     required this.createdAt,
     required this.updatedAt,
     this.relevance,
@@ -261,6 +264,7 @@ class _$MomentSearchResponseDto extends MomentSearchResponseDto {
         tipTotal == other.tipTotal &&
         viewerLiked == other.viewerLiked &&
         viewerBookmarked == other.viewerBookmarked &&
+        canInteract == other.canInteract &&
         createdAt == other.createdAt &&
         updatedAt == other.updatedAt &&
         relevance == other.relevance;
@@ -284,6 +288,7 @@ class _$MomentSearchResponseDto extends MomentSearchResponseDto {
     _$hash = $jc(_$hash, tipTotal.hashCode);
     _$hash = $jc(_$hash, viewerLiked.hashCode);
     _$hash = $jc(_$hash, viewerBookmarked.hashCode);
+    _$hash = $jc(_$hash, canInteract.hashCode);
     _$hash = $jc(_$hash, createdAt.hashCode);
     _$hash = $jc(_$hash, updatedAt.hashCode);
     _$hash = $jc(_$hash, relevance.hashCode);
@@ -309,6 +314,7 @@ class _$MomentSearchResponseDto extends MomentSearchResponseDto {
           ..add('tipTotal', tipTotal)
           ..add('viewerLiked', viewerLiked)
           ..add('viewerBookmarked', viewerBookmarked)
+          ..add('canInteract', canInteract)
           ..add('createdAt', createdAt)
           ..add('updatedAt', updatedAt)
           ..add('relevance', relevance))
@@ -391,6 +397,10 @@ class MomentSearchResponseDtoBuilder
   set viewerBookmarked(bool? viewerBookmarked) =>
       _$this._viewerBookmarked = viewerBookmarked;
 
+  bool? _canInteract;
+  bool? get canInteract => _$this._canInteract;
+  set canInteract(bool? canInteract) => _$this._canInteract = canInteract;
+
   DateTime? _createdAt;
   DateTime? get createdAt => _$this._createdAt;
   set createdAt(DateTime? createdAt) => _$this._createdAt = createdAt;
@@ -425,6 +435,7 @@ class MomentSearchResponseDtoBuilder
       _tipTotal = $v.tipTotal;
       _viewerLiked = $v.viewerLiked;
       _viewerBookmarked = $v.viewerBookmarked;
+      _canInteract = $v.canInteract;
       _createdAt = $v.createdAt;
       _updatedAt = $v.updatedAt;
       _relevance = $v.relevance;
@@ -519,6 +530,7 @@ class MomentSearchResponseDtoBuilder
               r'MomentSearchResponseDto',
               'viewerBookmarked',
             ),
+            canInteract: canInteract,
             createdAt: BuiltValueNullFieldError.checkNotNull(
               createdAt,
               r'MomentSearchResponseDto',